Remove incorrect quotation marks around Space and Enter
Standard English rules do not place quotation marks around mentions of keyboard keys. I removed them in English - and in the other translations to be consistent, although I don't legitimately know the rules for German or Polish, etc.

I'd changed this in my personally-installed version, but thought I would/should submit it as an official change in the master project. I'm somewhat of a grammar nerd.